refactor: tag, main-page components
