+++ frontend/src/components/admin/companies/tags/modal/tag-list.tsx
@@ -1,7 +1,7 @@
 import { memo } from 'react';
 import type { UseTagList } from 'hooks/common';
 
-import Tag from '../common/tag';
+import Tag from '../tag-item/tag-item';
 
 type Props = {
   tagList: UseTagList['list'];
